Update Karmanode Guide Intro
This begins the updated karmanode install guild and adds javascript for copy and paste for the rest of the guide.

This part of the guide is for desktop viewing only. A mobile viewport guide will be made the "old school" way but should not add to much load time to the page as it will be mostly text.

Note: downloadable pdf guide should be added for download and a video guild needs to be created to help the visual learners :)